+++ src/main.js
@@ -1,13 +1,15 @@
 /* TODO:
  * Configuration change watching
- * Bundling + re-organising
  * Clear LSP configuration before setting new config
+ * LSP gets confused after save
+ * No way to automatically add imports through code actions?
  */
 
 const RomeLint = require('./RomeLint.js');
 const RomeFormat = require('./RomeFormat.js');
 const notify = require('./notify.js');
 const Config = require('./config.js');
+const getRomePath = require('./path.js');
 
 let server = null;
 let formatter = null;
@@ -42,6 +44,17 @@ exports.activate = async function () {
             }
         });
 
+        await Config.registerOnChanged('be.aben.rome-enable-debug', async (value) => {
+            if (server) {
+                const restart = !!value !== server.debug;
+                server.debug = !!value;
+                restart && server.restart();
+            }
+            if (formatter) {
+                formatter.debug = !!value;
+            }
+        });
+
         nova.commands.register('be.aben.rome.formatDocument', async (obj) => {
             /* either a workspace or a text editor was passed */
             let editor = TextEditor.isTextEditor(obj) ? obj : obj.activeTextEditor;
@@ -63,75 +76,4 @@ exports.deactivate = function () {
     }
     Config.dispose();
 };
-
-async function getArch() {
-    return new Promise((res) => {
-        var process = new Process('/usr/bin/uname', { args: ['-m'] });
-        process.onStdout((line) => {
-            cpu = line.trim();
-            res(cpu);
-        });
-        process.start();
-    });
-}
-
-async function checkLocalPath() {
-    const arch = await getArch();
-    return new Promise((res) => {
-        /* Try to find local rome path in repo */
-        const localPath = `${nova.workspace.path}/node_modules/@rometools/cli-darwin-${arch}/rome`;
-        try {
-            const process = new Process('/usr/bin/env', {
-                args: ['stat', localPath],
-            });
-            process.onDidExit((value) => {
-                if (value) {
-                    res(null);
-                } else {
-                    res(localPath);
-                }
-            });
-            process.start();
-        } catch (err) {
-            res(null);
-        }
-    });
-}
-
-function checkGlobalPath() {
-    return new Promise((res) => {
-        /* Try to find global path */
-        try {
-            const process = new Process('/usr/bin/env', {
-                args: ['type', '-a', 'rome'],
-            });
-            let result = null;
-            process.onStdout((value) => {
-                const m = value.match(/rome is (\S+)/);
-                if (m?.length === 2) {
-                    result = m[1];
-                }
-            });
-            process.onDidExit((value) => {
-                if (result !== null) {
-                    res(result);
-                } else {
-                    res(null);
-                }
-            });
-            process.start();
-        } catch (err) {
-            res(null);
-        }
-    });
-}
-
-async function getRomePath() {
-    let path = await checkLocalPath();
-    if (path !== null) {
-        return path;
-    }
-    path = await checkGlobalPath();
-    return path;
-}